Add files via upload
Load/save files now mostly *should* work but please file any needed bug reports

+++ globals.h
@@ -12,7 +12,7 @@
 typedef struct objects {
 	char name[30];
 	int weight;
-	int id;	
+	short unsigned int id;	
 	struct objects *location; //set location to NULL to destroy an object
 	bool can_move;
 	bool been_here;
